% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_odf}
\alias{data_odf}
\title{data_odf}
\format{
A data frame with 20 rows and 7 variables:
\describe{
\item{bap87}{Current Health.}
\item{bap9201}{Hours of sleep, normal workday.}
\item{bap9001}{Pressed For Time Last 4 Weeks.}
\item{bap9002}{Run-down, Melancholy Last 4 Weeks.}
\item{bap9003}{Well-balanced Last 4 Weeks.}
\item{bap96}{Height.}
\item{name}{Firstname.}
}
}
\source{
\url{https://github.com/opendataformat/Specification/tree/main/Example}
}
\usage{
data_odf
}
\description{
example data with attributes specified for the Open Data Format.
}
\keyword{datasets}
